log most content resolver errors also in json backup object
Most of the exceptions are caught in the dumper class and skipped.
This to avoid unknown or even unfixable errors to prevent at least some
parts of the dump to be successful.

However, we want to be aware if errors happened, and some messages are
missing from the backup. So we record them (mostly) in the JSON object.
As a result, we might have a partial backup with info on what failed,
and hopefully info on how to fix it.