converters/maildir: have virtual root for each conversation
Having the first SMS/MMS as the root of a thread prevents continuing the
same thread if the root disappears. Instead, we create a virtual root,
without a message attached, it's just an ID referred to by all other
messages in the thread.

Since a thread consists in messages with the same group of recipients,
the thread-id (or message-id of the root) will be the hash of the
recipient group. It's deterministic as recipients cannot change.

Threading can be disabled with an option.