converters/html: look for "addresses" key before "address" key
BreakTheSilence converter uses columns from Silence database and the address
might be set though it's a conversation with more than 2 peers, so "addresses"
is also set, unlike Epistolaire's supposed output, where "address" would not
be set.
It shouldn't break anything, "addresses" is supposed to be set less often
than "address" so we can take it first.
It will prevent from writing multiple conversations with a shared peer to the
same file, and instead it will generate one file per list-of-recipients.